+++ FlowofEnglish_Backend/src/main/java/com/FlowofEnglish/service/UserAttemptsService.java
@@ -6,8 +6,8 @@ import java.util.Optional;
 
 public interface UserAttemptsService {
     List<UserAttempts> getAllUserAttempts();
-    Optional<UserAttempts> getUserAttemptById(int userAttemptId);
+    Optional<UserAttempts> getUserAttemptById(Long userAttemptId);
     UserAttempts createUserAttempt(UserAttempts userAttempt);
-    UserAttempts updateUserAttempt(int userAttemptId, UserAttempts userAttempt);
-    void deleteUserAttempt(int userAttemptId);
+    UserAttempts updateUserAttempt(Long userAttemptId, UserAttempts userAttempt);
+    void deleteUserAttempt(Long userAttemptId);
 }
